Replace calls to `container` with `registry`
`Ember.Registry` has not yet landed on `stable`, but floods the console
with deprecation warnings for `beta` and `canary`.

This PR replaces all calls of `env.container.register` with
`env.registry.register` throughout the test suite.

When the Registry is available, it will handle registration. Otherwise,
it will defer to the Container.