move factories off constructor and delegate to the container for lookups
This commit changes the way factories are located by Ember Data.
Previously, factories were stuffed onto the constructor of a
DS.Model subclass.

This was problematic for a few reasons:

- Classes for relationships defined via hasMany(Object) (instead of
  hasMany('string-name') were different than what was in the container.
  This caused several issues for Ember CLI and Globals Mode users alike.
- Inconsistency with the rest of the way lookups work in Ember.
- Zalgo

Now, relationship information can be accessed off the "type" in the
relationships "meta" hash. If you need to lookup a factory in your app,
you need to use this.store.modelFor, or this.store.modelFactoryFor if
you prefer to check that the model exists without throwing an exception.
This change makes all model factory lookups go through the container.
This should hopefully kill `Ember.MODEL_FACTORY_INJECTIONS = true;`.