Removed 'application' Adapter Lookup in adapterFor.
With a previous commit removing the default value of `-rest` against `store.adapter` this reference to the `application` adapter is no longer required. the `application` adapter will still be looked up second under default scenarios (in `defaultAdapter`).

In scenarios where the user provides a value to `store.adapter`, the `application` adapter will be looked up third - this logic is contained within `defaultAdapter`.