serializerFor Consistency
- Changed Adapter Lookup
Previously a call to `adapterFor` was used - then the `defaultSerializer` property of the returned adapter would be looked up.
This has potential to return a 'default' adapter such as the `application` adapter or `-rest` adapter.
It doesn't seem right that in a function whose purpose is to return a serializer, a default adapter (via its `defaultSerializer` property) would get preference over the default serializer (`-default`)
So this has been changed to call `lookupAdapter` - which will only ever return the adapter for the provided `type` and never a 'default' adapter.

- Implemented defaultSerializer in serializerFor
This also helps with the above situation by allowing the `application` serializer and `-default` serializer to be returned. It also makes things consistent with regards to adapterFor

- Removed reference to the `application` serializer in `serializerFor`
This is now covered in the `defaultSerializer`.