Refactor internals to use Javascript objects for internals instead of DS.Model
This commit adds an `InternalModel` class that is now used everywhere inside ED
to represent models. This class is a fast Javascript object that contains all the data
that we know for a particular record. At the ED/App code boundaries, such as responses
to `find`, models being set/pushed to `belongsTo`/`hasMany` we convert between
internalModels and DS.Models.

This should be a huge performance win, because we now lazily create DS.Models which are
pretty slow to instantiate.

We need to wait for the new serializer/push refactor in order to use a `push` that does
not immediately materialize a record to get further perf gains. For now most of perf gains
if for foreign keys.