use dasherized model names everywhere / change typeKey -> modelName
Previously, we allowed container keys to be loosy-goosy and normalized
them in several places. This commit consolidates those efforts into
one place `normalizeTypeKey`.

typeKey is now referred to as modelName on snapshot instances and Model
classes. typeKey remains on both of these but is deprecated.

modelName is now a dasherized string, where previously it was
camelCased.