Remove logic for camelCase-ing error keys in ActiveModelAdapter
Since c947e254ac6df93bd8b1b836e526fbd7b648f8ba the keys of the error
object are camelCased in the adapter, which is actually a serializer
concern. The JSONSerializer, from which the ActiveModelSerializer
extends, does a normalization of the errors object in the
`extractErrors` hook (see http://git.io/vJ0A6 & http://git.io/vJ0AS) and
this hook gets called by the store when the adapter rejects (see
http://git.io/vJ0AI), the camelCase can be removed from inside the
adapter.

A test for the correct normalization of the errors object is added to
the ActiveModelSerializer test suite.