Pass through the parsed response directly again
The changes in e928355f doesn't make the behaviour as backwards compatible as
we would like. The changes would break existing systems that do depend on a
snake case errors hash as we've seen in #3067.

For now restore the original behaviour so that we can make a release. We should
make a decision at some point whether to use the documented interface for
`InvalidError` or to change the interface.

Part of the issue here is that we don't have an integration test to qualify that
the errors hash gets read correctly and camelized. There's been some confusion
as to whether or not ember data actually does this.