fix bower publishing
From the Travis CI docs:

http://docs.travis-ci.com/user/ci-environment/

For builds not triggered by a pull request this is the name of the
branch currently being built; whereas for builds triggered by a pull
request this is the name of the branch targeted by the pull request (in
many cases this will be master).

This was building for branches not intended to be published because they
were opened as a pull request against master. As Ember Data core
contributors frequently work off the main repository this is an unwanted
side effect.