**In reference to [json-api/json-api#804](https://github.com/json-api/json-api/issues/804) marked for JSON-API 1.1-beta milestone**
With the current serializer, the way that EmberData POSTs
`belongsTo` relationships, it is impossible for the server to know what the
application wants to name the inverse relationship when sending the object with
the `hasMany` collection.

This patch adds an `inverse` attribute to the relationship payload when creating
an object with a `belongsTo` so that the referenced object
with the `hasMany` knows what to call the relationship.

In the example below, POSTing a new PingBack will set the `inverse` to
`incoming-links` which tells `post#123` what to name the
relationship when returning the collection of `ping-backs`.

The patch adds the `inverseBelongsToKey()` method to the JsonApiSerializer
which find the attribute name where the model for a `hasMany` matches the
model of the `belongsTo` relationship. It also respects
`{inverse: 'otherName'}` defined with `DS.attr()`.

```javascript
Post = DS.Model.extend({
  title: DS.attr('string'),
  incomingLinks: DS.hasMany('ping-back', { async: true })
});

PingBack = DS.Model.extend({
  post: DS.belongsTo('post', { async: true }),
  url:  DS.attr('string')
});
```

```javascript
// POST /ping-backs
{
  "data": {
    "type": "ping-backs",
    "attributes": {
      "url": "http://www.example.com/"
    },
    "relationships": {
      "post": {
        "inverse": "incoming-links", // <-- Name of inverse relationship is now defined as `incoming-links`
        "data": { "type": "post", "id": "123" }
      }
    }
  }
}

// GET /posts/123
{
  "data": {
    "type": "posts",
    "id": "123",
    "relationships": {
      "incoming-links": { // <--- Name of relationship can be known via 'inverse' attribute on relationship.
        "data": [
          { "type": "ping-backs", "id": "456" },
          { "type": "ping-backs", "id": "789" }
        ]
      }
    }
  }
}
```