Add ember-cli-dependency-checker.
Since we are using ember-cli to build ember-data we can take advantage
of the work done in ember-cli-dependency-checker to ensure that the deps
in `node_modules` and `bower_components` actually match what is
specified in the `package.json`/`bower.json`.

With this added, running `ember` commands will error if matching
versions are not found.