[BUGFIX beta] Fix usage of registry for 2.1.0+.
Using `registry.resolve` is generally private API, and `container.lookupFactory`
provides roughly the same result.

We still need the registry to actually register the model we create, so
I added a small matrix inline (this was the only usage of accessing the
registry off of `this.container` that I found in the codebase).