Remove references to findQuery
`findQuery` was deprecated back in 1.0 and removed for 2.0

A test file was also removed that was referencing `findQuery` as well
as `RestAdapter#find` (which was also deprecated and removed).

In addition to referencing code that was no longer present, this test
file was missing the `-test` in its name and has not been running as
part of the build. It did not pass before or after these changes.