[BUGFIX beta] serializePolymorphicType uses keyForPolymorphicType
Currently `keyForAttribute` is used to get the key under which the type
of a polymorphic record is serialized. This is not correct, as the new
`keyForPolymorphicType` hook should be used.

This commit uses the new hook and falls back to the previous vesion, if
the key generated via the old method (keyForAttribute) is different to
the new version and the `keyForPolymorphicType` has not been
overwritten. A deprecation warning is logged if this is the case.