`Store._find` asserts `adapterPayload` not empty
Currently, [Store._find][find] only asserts that the `adapterPayload` is
non-null (i.e. `{}` is a valid payload).

This commit forces the `adapterPayload` to be non-empty.

When TDDing with Mirage, `adapterPayload` is often `{}`, which sneaks
through the asserts and ends up as a non-descriptive:

```
Assertion Failed: You must include an 'id' for undefined in an object
passed to 'push'
```

Which is triggered in [Store._pushInternalModel][_pushInternalModel],
which is a few method invocations away from the source of the issue.

[_pushInternalModel]: https://github.com/emberjs/data/blob/f1ccad8ab9356dd51f44d63585c27d8bb4ec9c3a/packages/ember-data/lib/system/store.js#L1709
[find]: https://github.com/emberjs/data/blob/f1ccad8ab9356dd51f44d63585c27d8bb4ec9c3a/packages/ember-data/lib/system/store/finders.js#L27