Re-add possibility to enable feature flags within Browser
Optional feature flags can now be enabled via a checkbox, rendered at
the top of the test page. The state of the checkbox is set as query
parameter in the URL.

An in-repo-addon ensures that the ENABLE_OPTIONAL_FEATURES flag works
correctly, when the tests are run inside a browser: the workaround using
the in-repo-addon is needed, since the flag needs to be set before
Ember.js is loaded, but there is currently no other way.