[BUGFIX beta] Use `bower` version of ember-data if present.
In order to keep backwards compat, we must honor the version of
ember-data that is present in `bower.json` (even though it is not
needed). This is due to the fact that many people update the
`package.json` and `bower.json` versions of Ember Data completely
independently.

If `bower.json` includes `ember-data` a warning will be issued, and that
version will be used.

(cherry picked from commit 446d65f7ebbb6b34a3dc50c24a5d992889d581a1)