Change test for asserting that model is available
Currently `this._hasModelFor` returns the specific model, though - as
the name implies - a boolean should be returned.

---

Context: this change is especially needed so the assertion helper from
ember-dev works correctly: this helper uses the checkTest function in
test-helper/utils.js https://git.io/vuZK3. If a class is passed to the
assertion, like in `Ember.assert("should exists", MyEmberClass)`, it is
treated as a function and invoked within `checkTest`. This throws an
error since ember classes are not function which can be invoked.