[BUGFIX release] don't load "app" code when loading globals files
The code that gets compiled for the globals build is in a closure with
its own instance of `loader.js`, so the entries for ember-data don't end
up in ember-cli's instance of `loader.js`. The ember-data code does
include shims that end up in the app's instance of loader.js, so we
exclude the app code from running in this case.

(cherry picked from commit 1a861bb19e96f5e01930e8d5651abf175e989379)