[CLEANUP] Remove shim for Backburner.join
This shim was added a long while back when we could not rely on the
version of Backburner that Ember was using to have the `join()`
function.

Any currently supported version of Ember (2.x) should have this now.

Similarly, any currently supported (2.x) version of Ember also has
`Ember._Backburner` available.

Shim was added in:
https://github.com/emberjs/data/pull/2608

`Backburner.join()` was added in:
https://github.com/ebryn/backburner.js/pull/119