[BUGFIX release] Fix RESTAdapter.buildQuery without a snapshot
Before introducing a feature ds-finder-include it was possible to run
findRecord on an adapter without passing a snapshot. After introducing
the feature it's no longer possible even if the feature is disabled,
because some of the code was outsief of the flag check.

This commit fixes the code to work like before introducing the
ds-finder-include change.