Split test for rejected records in store.fetchRecord
This test contains assertions for correctly rejected records in
coalesced requests AND an assertion that a warning is logged.

This commit splits the test into two: the first tests the correct
behavior for rejected records and the second tests that the warning for
missing records is logged.

The splitting is necessary since in an upcoming commit the tests are run
in production too: since in that environment tests which check for
assertions are stripped, currently we would lose this test. That's why
this test is split beforehand so the behavior is tested in production
build as well.