Run tests in production environment
This adds a `testInDebug` helper, which skips tests when run in
production. This helper either runs the tests normally via QUnit.test if
we are not in a production environment. If the test is executed within
production however, QUnit.skip is used instead so the test is skipped.

This allows us to skip tests which don't make sense in the production
environment: this are tests which check for warnings and assertions
thrown within code. Since those assertions/warnings are stripped from
production code, it doesn't make sense to run those tests in that
environment.