[BUGFIX release] Revert pr #3864
The origial intent was for a queryRecord with no response to reject the promise so it would act like a `findRecord` with a 404. This change introduced a regression that broke existing apps so it is going to be reverted.

Closes #4219

(cherry picked from commit 274d7b1622c578bc47ea56c7bb86dd8f0e6f03db)