[BUGFIX beta] fix isUpdating for AdapterPopulatedRecordArray#update()
The `isUpdating` flag is not set to `true` when the `update()` method on
a `DS.AdapterPopulatedRecordArray` is invoked. As with the flag on
`DS.RecordArray`, this should be `true` until the update is finished and
the array contains the most reset result for the query, fetched from the
adapter.

(cherry picked from commit 89ce6cde4070b68606cef4fc66364b6b5462e9cd)