[FEATURE ds-payload-hooks] Add hooks to map type in payload to modelName
The `modelNameFromPayloadKey` and `payloadKeyFromModelName` hooks on the
serializer allow to customize the mapping between the key of a JSON and
the corresponding name of the model. Consider the following payload:

```json
{
  "blog/post": {
    "id": 1
  }
}
```

To map the "blog/post" key to the `post` model, the
`modelNameFromPayloadKey` hook is used:

```js
serializer.modelNameFromPayloadKey("blog/post")
```

---

There are some issues with the current code base, as the
`modelNameFromPayloadKey` and `payloadKeyFromModelName` hooks are also
used to map the "value" of a type from / to the payload. Consider the
following payload of a JSON-API document:

```json
{
  "data": {
    "id": 1,
    "type": "API:V1::User"
  }
}
```

To map the namespaced type to the model name ember-data is using,
currently the `modelNameFromPayloadKey` is used:

```js
serializer.modelNameFromPayloadKey("API::V1::User")
```

Now this gets complicated if your API responds with the following
payload (using custom key and custom types for polymorphic records for
example):

```json
{
  "blog/post": {
    "id": 1,
    "user": 2,
    "userType": "API:V1::Administrator"
  }
}
```

Now the `modelNameFromPayloadKey` is invoked for both:

```js
serializer.modelNameFromPayloadKey("blog/post")
serializer.modelNameFromPayloadKey("API::V1::Administrator")
```

This means that the logic within the hook would get complicated.

---

As the name suggests, the method should only be used to map the key, not
the value. This commit adds `modelNameFromPayloadType` and
`payloadTypeFromModelName` hooks and uses them during normalization and
serialization.

This commit maintains the old behavior using, if it is used and logs a
deprecation warning to move to the new hooks.