[BUGFIX release] Ensure feature flag stripping works for all builds.
Prior to this, the feature flags would only be processed for
`production` builds (from within an ember-cli app). This is incorrect,
they should be processed for all builds (for every environment) so that
any features explicitly enabled or disabled in the version of Ember
data being used are stripped properly.