[BUGFIX release] Only setup babel options once.
`included` is called for each instance of `EmberApp` created
within the consuming applicationm, which means that we would be adding
the same plugins if the consuming application happened to be creating
two `EmberApp` instances.

For example, ember-cli-fastboot currently uses this
technique to build both for fastboot AND the normal build. I know of a
few other applications that use similar techniques to build two versions
of the same `EmberApp` (i.e. a mobile version and a desktop one).

This allows us to ensure that `this.options.babel` is only setup once
(regardless of how many times `included` is being called) during init.