Revert "[CLEANUP ds-serialize-ids-and-types]"
This reverts commit 87e6e60e9b6266b27eceae17891839cf2b7a1787.

The cleanup for the ds-serialize-ids-and-types feature is made too
eagerly, since it hasn't been enabled in beta yet. This commit reverts
the cleanup so the functionality is still behind the (enabled) feature
flag. By this it can be tested via the beta channel.

Once the feature is go'ed in a release channel, this can be finally be
cleaned.