[BUGFIX beta] Ensure record array length is reset during willDestroy.
In Ember 2.10+ objects that are in the middle of being destroyed no
longer notify property changes on itself (or other destroying objects)
to avoid wasted work.  This means that `set(this, 'content',
undefined)` isn't enough to ensure that `Enumerable` mixin will
`notifyPropertyChange(this, 'length')` (since its observers are not
fired).

This change also sets `length` during cleanup so it properly resets to
`0`.