[DOC] Update descriptions of findRecord() and findAll()
This commit adds documentation on the use of the `include` query parameter in `DS.Store` methods `findRecord()` and `findAll()`. This feature allows records of models related to those requested to be retrieved from JSON API compliant server in a single request. This capability was added in https://github.com/emberjs/data/pull/3976 but is not currently documented.