[FEATURE ds-overhaul-references] only allow JSON-API and records
This commit updates the existing `ds-overhaul-references` feature.

It tackles some inconsistencies within `push()` on references, so the
following arguments will be accepted for `push()`:

- `BelongsToReference#push(arg)`
  - `arg` is an instance of `DS.Model`
  - `arg` is a JSON-API document for a single resource object
- `HasManyReference#push(arg)`
  - `arg` is an array of `DS.Model`s
  - `arg` is a JSON-API document for an array of resource objects

`Reference#push(object)` also sets the corresponding `meta` and
`links` on the underlying relationships.

---

If the passed argument to `HasManyReference#push()` has one of the
following structures, a deprecation will be logged:

- `{ data: [{ data: { … } }, { data: { … } }] }`
-`[{ data: { … } }, { data: { … } }]`