Avoid mutating model factory in _modelForMixin.
With the upcoming `factory-for` work it is likely that mutating the
result of `modelFactoryFor` will throw an error.  This is a minor
refactoring to avoid needing to mutate the return value (since we
can simply do the work before registering).

(cherry picked from commit 9aa6f091dd8a0f34a9ec767123db693d642a65dd)