[BUGFIX beta] Ensure Engines can boot without error.
The argument passed into `instance-initializers` is _either_ an
`Ember.ApplicationInstance` instance _or_ a `Ember.EngineInstance`
instance. The `.application` property is only present on
`Ember.ApplicationInstance`'s.

This change adds some inline comments around the various conditionals
and swaps the logic to use `instance.base` when present and avoid
erroring if neither `.base` or `.application` is present.