[BUGFIX beta] Implement `cacheKeyForTree` hook.
This hook was added in [ember-cli/rfcs#90](https://github.com/ember-cli/rfcs/blob/master/complete/0090-addon-tree-caching.md),
and essentially guarantees that a given tree returned by an addon is
considered stable (or not).

In the case of Ember Data, we do not have different tree output
based on our parent (project or addon).

Specifically, this implementation allows both an app _and_ a lazy engine to
depend on `ember-data` without duplicating the `ember-data` assets in both
the `assets/vendor.js` and `engine-dist/<engine-name>/assets/engine-vendor.js`.