Ensure CI has latest yarn.
The bizarre failures on CI are related to the super old Yarn version
in use on the Trusty arch (0.17.5), and seem to randomly fail based
on which CI image a build runs on (super frustratingly :P ).

This updates our Travis config to ensure the latest yarn is installed
(and also updates the mechansim we use to get phantom so that it works
across both Trusty and Precise images).