[BUGFIX beta] Ignore included resources without that don't have a corresponding ember-data model
When the JSON API serializer encounters a resource with a type that doesn't resolve
to one of the application's Ember Data models, it replaces the object with `null` and
emits a warning. Prior to this commit, if that `null` was one of the sideloaded
(i.e. `included`) resources, the store would throw trying to get the `null`'s type.
This commit makes the serializer drops any nulls out of the list of included resources
to avoid this problem.

To summarize, if your API attempts to include resources that you don't have models for,
a warning will be emitted and they will be ignored.

Fixes #4568