[BUGFIX beta] Avoid cache related warnings during builds.
broccoli-babel-transpiler (which we use by way of ember-cli-babel)
requires that all plugins have a `.baseDir` function which
broccoli-babel-transpiler uses to ensure that it can properly invalidate
the cached output of each transpiled file. If `.baseDir` is not present
broccoli-babel-transpiler does the only "safe" thing and completely opts
out of this caching system and prints a very loud warning message.

This commit ensures that the few custom plugins that we are using are
properly setup with a `.baseDir`.