[BUGFIX beta] Use native Map if present.
When native `Map` is present, use it. Otherwise fallback to `Ember.Map`.

This positions ember-data for the eventual deprecation and dropping
of `Ember.Map` and `Ember.MapWithDefault` in Ember itself (and using
native classes should generally be better for us long term than shipping
custom _slightly_ different versions).

Note there are a few specific differences between native `Map` (and the
`MapWithDefault` implementation that extends from the native `Map` when
possible) and Ember's `Ember.Map` that we should be aware of:

* `Ember.Map` has custom `copy` and `isEmpty` methods which are not
present in native `Map`
* `Ember.Map` adds a static `create` method (which simply instantiates
itself with `new Ember.Map()`)
* `Ember.Map` does not accept constructor arguments
* `Ember.Map` does not have:
  * `@@species`
  * `@@iterator`
  * `entries`
  * `values`