[BUGFIX beta] Remove non-functional (error prone) manual CP caching.
The previous implementation was intending to cache relationships when in
the production environment, but avoid caching them in development.
Unfortunately, this code was factored in a way that relied on mutating
internal private state of the underlying `Ember.ComputedProperty`
instance when `Ember.testing` was true (which was used as a way to
determine prod vs non-prod).

When this code was introduced (2014-12-02), setting `_cacheable` on a
`ComputedProperty` instance would have worked as intended, however a
refactor in Ember (ironically one day prior, but on Ember's canary
channel) deprecated using `_cacheable` or `.cacheable()` (in favor of
opting out of caching via `.volatile()`).

Since this has not worked since Ember 2.0.0, and it has not been an
issue, this commit completely removes the manual caching / volatile
swapping between prod/dev/testing environments.

(cherry picked from commit 18a0534c5648828f3360f1eb4079161dafd13737)